requirements.txt
chromadb
fastapi
pydantic
python-dotenv
